Move deletes to sqlstore and run them in a transaction
